requirements.txt
Flask == 3.1.0
psycopg2-binary == 2.9.10
# urllib3 == 2.6.1
requests == 2.32.3
Flask-SQLAlchemy == 3.1.1
Flask-RESTful == 0.3.10